+++ frontend/providers/dbprovider/src/pages/db/detail/index.tsx
@@ -22,6 +22,7 @@ import ErrorLog from '@/pages/db/detail/components/ErrorLog';
 import MyIcon from '@/components/Icon';
 import { BackupSupportedDBTypeList } from '@/constants/db';
 import DataImport from './components/DataImport';
+import { getDBByName } from '@/api/db';
 
 enum TabEnum {
   pod = 'pod',
@@ -254,9 +255,16 @@ const AppDetail = ({
 export default AppDetail;
 
 export async function getServerSideProps(context: any) {
+  let dbType = context.query?.dbType || '';
+
   const dbName = context.query?.name || '';
-  const dbType = context.query?.dbType || '';
   const listType = context.query?.listType || TabEnum.Overview;
+
+  if (!dbType) {
+    const baseURL = `http://${process.env.HOSTNAME || 'localhost'}:${process.env.PORT || 3000}`;
+    const db = await getDBByName(dbName, { baseURL });
+    dbType = db.dbType;
+  }
 
   return {
     props: { ...(await serviceSideProps(context)), dbName, listType, dbType }