+++ frontend/providers/template/src/pages/_document.tsx
@@ -1,15 +1,17 @@
 import { Html, Head, Main, NextScript } from 'next/document';
 
 export default function Document() {
+  const brandName = process.env.NEXT_PUBLIC_BRAND_NAME;
+
   return (
     <Html lang="en">
       <Head>
-        <meta name="application-name" content="Sealos Desktop App Demo" />
+        <meta name="application-name" content={`${brandName} Desktop App Demo`} />
         <meta name="apple-mobile-web-app-capable" content="yes" />
         <meta name="mobile-web-app-capable" content="yes" />
         <meta name="apple-mobile-web-app-status-bar-style" content="default" />
-        <meta name="apple-mobile-web-app-title" content="Sealos" />
-        <meta name="description" content="sealos cloud" />
+        <meta name="apple-mobile-web-app-title" content={brandName} />
+        <meta name="description" content={`${brandName} cloud`} />
         <meta name="format-detection" content="telephone=no" />
       </Head>
       <body>